Add files via upload
It is a console application that is launched from the command line using Java. The server and clients can run on different computers in the same network, using Web Socket programming. There is a client connected to a server and they can chat with each other. The application consists of two parts: server and client. Each part can run independently on separate computers.